requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pdfplumber
pdf2image
pillow
easyocr
ddgs
newsapi-python
requests
python-dotenv
google-genai
groq